requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.7
django-appconf==1.0.6
django-compressor==4.5.1
exceptiongroup==1.2.2
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
oauthlib==3.2.2
openai==1.37.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
rcssmin==1.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
